# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-commerce shop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Customer {
  id: number;
  phone: string;
  name?: string | null;
  avatarUrl?: string | null;
  language?: string | null;
  telegramId?: string | null;
  createdAt: string;
}

export interface CustomerSearchBody {
  phone: string;
}

export interface CustomerSearchResult {
  exists: boolean;
  customer?: Customer | null;
}

export interface CustomerLoginBody {
  phone: string;
  name?: string;
}

export interface UpdateProfileBody {
  name?: string;
  avatarUrl?: string;
  language?: string;
}

export interface Product {
  id: number;
  name: string;
  description?: string | null;
  price: number;
  oldPrice?: number | null;
  images: string[];
  categoryId?: number | null;
  categoryName?: string | null;
  inStock: boolean;
  isLiked: boolean;
  createdAt: string;
}

export interface ProductsListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface CreateProductBody {
  name: string;
  description?: string;
  price: number;
  oldPrice?: number;
  images: string[];
  categoryId?: number;
  inStock?: boolean;
}

export interface UpdateProductBody {
  name?: string;
  description?: string;
  price?: number;
  oldPrice?: number;
  images?: string[];
  categoryId?: number;
  inStock?: boolean;
}

export interface Category {
  id: number;
  name: string;
  imageUrl?: string | null;
  productCount: number;
}

export interface CreateCategoryBody {
  name: string;
  imageUrl?: string;
}

export interface CartItem {
  id: number;
  productId: number;
  quantity: number;
  product: Product;
}

export interface AddToCartBody {
  productId: number;
  quantity?: number;
}

export interface UpdateCartItemBody {
  quantity: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  new: "new",
  preparing: "preparing",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type OrderDeliveryMethod =
  (typeof OrderDeliveryMethod)[keyof typeof OrderDeliveryMethod];

export const OrderDeliveryMethod = {
  delivery: "delivery",
  pickup: "pickup",
} as const;

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  cash: "cash",
  card: "card",
  online: "online",
} as const;

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  productImage?: string | null;
  quantity: number;
  price: number;
}

export interface Order {
  id: number;
  customerId: number;
  customerName?: string | null;
  customerPhone?: string | null;
  status: OrderStatus;
  deliveryMethod: OrderDeliveryMethod;
  paymentMethod: OrderPaymentMethod;
  address?: string | null;
  note?: string | null;
  totalPrice: number;
  deliveryFee: number;
  items: OrderItem[];
  createdAt: string;
}

export type CreateOrderBodyDeliveryMethod =
  (typeof CreateOrderBodyDeliveryMethod)[keyof typeof CreateOrderBodyDeliveryMethod];

export const CreateOrderBodyDeliveryMethod = {
  delivery: "delivery",
  pickup: "pickup",
} as const;

export type CreateOrderBodyPaymentMethod =
  (typeof CreateOrderBodyPaymentMethod)[keyof typeof CreateOrderBodyPaymentMethod];

export const CreateOrderBodyPaymentMethod = {
  cash: "cash",
  card: "card",
  online: "online",
} as const;

export interface CreateOrderBody {
  deliveryMethod: CreateOrderBodyDeliveryMethod;
  paymentMethod: CreateOrderBodyPaymentMethod;
  address?: string;
  note?: string;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  new: "new",
  preparing: "preparing",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export interface ToggleLikeBody {
  productId: number;
}

export interface LikeResponse {
  liked: boolean;
}

export type MessageSenderType =
  (typeof MessageSenderType)[keyof typeof MessageSenderType];

export const MessageSenderType = {
  customer: "customer",
  admin: "admin",
} as const;

export interface Message {
  id: number;
  customerId: number;
  senderType: MessageSenderType;
  text: string;
  isRead: boolean;
  createdAt: string;
}

export type SendMessageBodySenderType =
  (typeof SendMessageBodySenderType)[keyof typeof SendMessageBodySenderType];

export const SendMessageBodySenderType = {
  customer: "customer",
  admin: "admin",
} as const;

export interface SendMessageBody {
  customerId?: number;
  text: string;
  senderType: SendMessageBodySenderType;
}

export type MarkReadBodySenderType =
  (typeof MarkReadBodySenderType)[keyof typeof MarkReadBodySenderType];

export const MarkReadBodySenderType = {
  customer: "customer",
  admin: "admin",
} as const;

export interface MarkReadBody {
  customerId: number;
  senderType: MarkReadBodySenderType;
}

export interface Banner {
  id: number;
  imageUrl: string;
  title?: string | null;
  link?: string | null;
  isActive: boolean;
  sortOrder: number;
  createdAt: string;
}

export interface CreateBannerBody {
  imageUrl: string;
  title?: string;
  link?: string;
  isActive?: boolean;
  sortOrder?: number;
}

export interface DeliverySettings {
  deliveryFee: number;
  freeDeliveryThreshold: number;
  estimatedMinutes: number;
}

export interface UpdateDeliveryBody {
  deliveryFee?: number;
  freeDeliveryThreshold?: number;
  estimatedMinutes?: number;
}

export interface SupportContact {
  phone: string;
  telegram?: string | null;
}

export interface UpdatePasswordBody {
  currentPassword: string;
  newPassword: string;
}

export interface AdminLoginBody {
  password: string;
}

export interface SendNotificationBody {
  message: string;
}

export interface DashboardStats {
  totalOrders: number;
  totalCustomers: number;
  totalRevenue: number;
  newOrders: number;
  preparingOrders: number;
  deliveredOrders: number;
  todayRevenue: number;
  todayOrders: number;
}

export interface OrderStatusCount {
  status: string;
  count: number;
}

export type ListProductsParams = {
  categoryId?: number;
  search?: string;
  page?: number;
  limit?: number;
};

export type ListOrdersParams = {
  status?: ListOrdersStatus;
  customerId?: number;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  new: "new",
  preparing: "preparing",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type ListMessagesParams = {
  customerId?: number;
};

export type TelegramWebhookBody = { [key: string]: unknown };