# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: E-commerce shop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers/search:
    post:
      operationId: searchCustomer
      tags: [auth]
      summary: Search customer by phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerSearchBody"
      responses:
        "200":
          description: Customer found or not
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSearchResult"
  /customers/login:
    post:
      operationId: loginCustomer
      tags: [auth]
      summary: Login or register customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerLoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current customer
      responses:
        "200":
          description: Current customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get cart items
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartItem"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
  /cart/{id}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove cart item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages for customer or admin
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/read:
    patch:
      operationId: markMessagesRead
      tags: [messages]
      summary: Mark messages as read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkReadBody"
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      summary: List banners
      responses:
        "200":
          description: Banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      summary: Create banner (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerBody"
      responses:
        "201":
          description: Created banner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /banners/{id}:
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    Customer:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        telegramId:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - phone
        - createdAt
    CustomerSearchBody:
      type: object
      properties:
        phone:
          type: string
      required:
        - phone
    CustomerSearchResult:
      type: object
      properties:
        exists:
          type: boolean
        customer:
          $ref: "#/components/schemas/Customer"
          nullable: true
      required:
        - exists
    CustomerLoginBody:
      type: object
      properties:
        phone:
          type: string
        name:
          type: string
      required:
        - phone
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: string
        language:
          type: string
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        oldPrice:
          type: number
          nullable: true
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        inStock:
          type: boolean
        isLiked:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - price
        - images
        - inStock
        - isLiked
        - createdAt
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: string
          nullable: true
        productCount:
          type: integer
      required:
        - id
        - name
        - productCount
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        imageUrl:
          type: string
      required:
        - name
    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
      required:
        - id
        - productId
        - quantity
        - product
    AddToCartBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - productId
    UpdateCartItemBody:
      type: object
      properties:
        quantity:
          type: integer
      required:
        - quantity
...
    Message:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        senderType:
          type: string
          enum: [customer, admin]
        text:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - customerId
        - senderType
        - text
        - isRead
        - createdAt
    SendMessageBody:
      type: object
      properties:
        customerId:
          type: integer
        text:
          type: string
        senderType:
          type: string
          enum: [customer, admin]
      required:
        - text
        - senderType
    MarkReadBody:
      type: object
      properties:
        customerId:
          type: integer
        senderType:
          type: string
          enum: [customer, admin]
      required:
        - customerId
        - senderType
    Banner:
      type: object
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        title:
          type: string
          nullable: true
        link:
          type: string
          nullable: true
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - imageUrl
        - isActive
        - sortOrder
        - createdAt
    CreateBannerBody:
      type: object
      properties:
        imageUrl:
          type: string
        title:
          type: string
        link:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
      required:
        - imageUrl
...